@@ -4,7 +4,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:http/http.dart' as http;
 import 'package:provider/provider.dart';
-import 'package:risu/components/alert_dialog.dart';
 import 'package:risu/components/appbar.dart';
 import 'package:risu/components/loader.dart';
 import 'package:risu/components/outlined_button.dart';
@@ -41,6 +40,7 @@ class NewTicketState extends State<NewTicketPage> {
           body: jsonEncode(<String, String>{
             'content': _content,
             'title': _subject,
+            'createdAt': DateTime.now().toString(),
           }));
       setState(() {
         _loaderManager.setIsLoading(false);
@@ -76,7 +76,7 @@ class NewTicketState extends State<NewTicketPage> {
         curveColor: context.select((ThemeProvider themeProvider) =>
             themeProvider.currentTheme.secondaryHeaderColor),
         showBackButton: false,
-        showLogo: true,
+        textTitle: AppLocalizations.of(context)!.newTicket,
       ),
       resizeToAvoidBottomInset: false,
       backgroundColor: context.select((ThemeProvider themeProvider) =>
@@ -92,16 +92,6 @@ class NewTicketState extends State<NewTicketPage> {
                     mainAxisAlignment: MainAxisAlignment.center,
                     children: [
                       const SizedBox(height: 30),
-                      Text(
-                        AppLocalizations.of(context)!.newTicket,
-                        style: TextStyle(
-                          fontWeight: FontWeight.bold,
-                          fontSize: 32,
-                          color: context.select((ThemeProvider themeProvider) =>
-                              themeProvider.currentTheme.primaryColor),
-                        ),
-                        textAlign: TextAlign.center,
-                      ),
                       const SizedBox(height: 32),
                       MyTextInput(
                         key: const Key('new-ticket-text_input-input-title'),
@@ -124,8 +114,8 @@ class NewTicketState extends State<NewTicketPage> {
                         key: const Key('new-ticket-button-send_message'),
                         onPressed: () async {
                           bool success = await createTicket();
-                          if (success) {
-                            Navigator.push(
+                          if (success && context.mounted) {
+                            Navigator.pop(
                               context,
                               MaterialPageRoute(
                                 builder: (context) => const ContactPage(),
