@@ -12,9 +12,9 @@ exports.sendAccountConfirmationEmail = (email, token) => {
     from: process.env.MAIL_ADDRESS,
     to: email,
     subject: 'Confirm your account',
-    text:
-      'Please follow the link to confirm your account: http://51.11.241.159:3000/api/mobile/auth/mailVerification?token=' +
-      token
+    text: "",
+    html: '<p>Please follow the link to confirm your account: <a href="http://risu.dns-dynamic.net:3000/api/mobile/auth/mailVerification?token=' +
+      token + '">here</a></p>',
   }
   try {
     transporter.sendMail(mailOptions)
