@@ -34,23 +34,47 @@ router.post('/signup', (req, res, next) => {
   )(req, res, next)
 })
 
-router.post('/login', (req, res, next) => {
+router.post('/login', jwtMiddleware.refreshTokenMiddleware, async (req, res, next) => {
   passport.authenticate(
     'login',
     { session: false },
-    (err, user, info) => {
+    async (err, user, info) =>  {
       if (err)
         throw new Error(err)
       if (user == false)
         return res.status(401).json(info)
 
-      const token = jwtMiddleware.generateToken(user.id)
+      const longTerm = req.body.longTerm || false;
+
+      const token = jwtMiddleware.generateToken(user.id, longTerm);
+
+      var refreshToken = '';
+      if (longTerm) {
+        refreshToken = jwtMiddleware.generateRefreshToken(user.id);
+        user = await userCtrl.updateUserRefreshToken(user.id, refreshToken)
+      } else {
+        user = await userCtrl.removeUserRefreshToken(user.id)
+      }
+
       return res.status(201).json({ user : user, token : token })
     }
   )(req, res, next)
 })
 
-router.get('/mailVerification', async (req, res) => {
+router.post('/login/refreshToken', jwtMiddleware.refreshTokenMiddleware, async (req, res) => {
+  const refreshToken = req.body.refreshToken;
+  if (!refreshToken || refreshToken == '') {
+    return res.status(401).send('No refresh token provided.')
+  }
+  const user = await userCtrl.findUserByRefreshToken(refreshToken)
+  if (!user) {
+    return res.status(401).send('No matching user found.')
+  }
+  const token = jwtMiddleware.generateToken(user.id)
+  return res.status(201).json({ user : user, token : token })
+})
+
+router.get('/mailVerification', jwtMiddleware.refreshTokenMiddleware, async (req, res) => {
   const token = req.query.token
   try {
     const decoded = jwt.decode(token, process.env.JWT_ACCESS_SECRET)
@@ -60,7 +84,6 @@ router.get('/mailVerification', async (req, res) => {
       'Email now successfully verified !\nYou can go back to login page.'
       )
   } catch (err) {
-    console.error(err.message)
     return res.status(401).send('No matching user found.')
   }
 })
