@@ -29,7 +29,7 @@ describe('POST /api/rent/article', () => {
             const res = await request('http://localhost:3000')
               .get('/api/mobile/article/listAll');
             expect(res.statusCode).toBe(200);
-            itemId = res.body[2].id;
+            itemId = res.body[1].id;
           }
         ],
         done
@@ -124,7 +124,7 @@ describe('POST /api/rent/article', () => {
         [
           async function () {
             const res = await request('http://localhost:3000')
-              .post('/api/mobile/rent/1/invoice')
+              .post('/api/mobile/rent/2/invoice')
               .set('Content-Type', 'application/json')
               .set('Accept', 'application/json')
               .set('Authorization', `Bearer ${authToken}`)
